qa/pmdas/dynamic/GNUmakefile: avoid somwe qa/956 failures
When running QA from a git tree (not the testsuite directory)
we try hard to use GNUmakefile.install in preference to
GNUmakefile so we're not dependent on the state of the git
tree.  But qa/956 exposes a corner case when the git tree
is pristine, then the dynamic PMDA (QA only special) fails
to Install because the Install script via pmdaproc.sh runs
a "make install" that uses GNUmakefile ... tart up GNUmakefile
to do the right thing when this happens.