qa/578: finally found cause of random failures
Since we introduced unix: as a channel to the local pmcd, it is no
longer safe to start pmcd with PMCD_PORT set to something non-standard
and assume the qa test is the only one talking to this pmcd.

Need to work harder to kill off all client connections to the test
pmcd instance ... it is a BIG deal for this test as it is checking
for pmcd fd leakage, and stray clients connection mess the test up.