src/pmcd/src/pmcd.c: just put the __pmOpenLog() call back where it was
Turns out it needs to be not only after the argument parsing, but also
after the pmcd request port is established (to avoid an accidentally
started second pmcd clobbering the first pmcd's log ... and yes there
is a qa test for this).

So the call is back where it originally was, but with a comment that
tries to explain why, so when my dementia gets worse I won't make the
same blunder again.

Sigh.