pmchart: small improvements to the vCPU view
Rearrange the order in which some guest metrics appear so
that when CPU and vCPU views used similtaneously the are a
direct visual match (i.e. more clearly showing which metrics
are divided up in the vCPU view when compared to CPU view).

Updated the repeated phrase in the heading for clarity too.