libpcp & man: rework interp logic arond <mark> records
Fix for https://bugzilla.redhat.com/show_bug.cgi?id=1296750

With this change, if a <mark> record is found between the previous
pmFetch() and the current pmFetch() we return "no values" for all
metrics.

The logic was already sound for cases where the requested time was
between the closest valid values and an adjacent <mark> record.

This change extends the logic it cases where the pmFetch() sample
interval is large compared to the pmlogger sample interval, and
there are lots of valid data records (and possibly one or more <mark>
records) between one pmFetch() time and the next.