Fix bug in interp mode record caching.
Records were cached based on
  1) metrics file pointer(!!!!)
  2) volume number
  3) offset in the metrics file

The problem with this is that it is possible for the metrics file
pointer to remain the same when a volume/archive change occurs. In this
case, if the previous volume is small enough, records from the new
volume may be determined to be cache hits for records in the cache
from the old volume.

This change uses the name of the archive instead of the metrics file
pointer, as is done elsewhere in libpcp when distinguishing archives.