containers: add pcp-pmwebd container to serve pcp-webapi requests.
This adds a pcp-pmwebd container. It needs a local pmcd service,
which can be provided either by a regular local pmcd on the host
_or_ by a running pcp-collector container. It listens on the default
pmwebd port (44323 by default). This can be changed by editing
pmwed.options in the container.

Host platforms that don't have an active avahi service will need
to add the new pmwebd -S flag in the pmwebd.options file in the
container to disable service advertising (see Nathan's follow-up
commit for this).

	modified:   build/containers/GNUmakefile
	new file:   build/containers/pcp-pmwebd/Dockerfile
	new file:   build/containers/pcp-pmwebd/GNUmakefile