pmrep: correctly report and use the --hostzone/-z option
Now that pmrep is more verbose in its reporting of timezones, we
are overridding some of the work done deep in libpcp getopt code
for setting up the host/archive-based timezone.  Needed a way to
access this from the python wrapper to solve this, and an update
to test 1069 (which is now deterministic in this aspect too).