python: extract strings with native encoding, always
Change python pmAtomValue dref() routine so that it will
pass appropriately encoded strings up to higher layers.
This avoids the need for pmExtractValue, fetchgroup, and
higher layers (like pmrep, where bug was found) having to
manually decode strings.  Complete the set of dref types.

This hides more of the py2/3 deltas from high-level APIs
and scripts, which is a universally Good Thing.

Test 1069 is updated to check pmcd.version string (esp.
ensuring its not prefixed with py3 b'...' bytes goop at
the time it is printed).