qa: add qa/847 to launch containers and test basic functionality
First draft of a qa test to actually launch pcp containers
and test basic functionality. The idea is to launch these
containers and have them check basic PCP functionality using
the local pmcd service. And then to stop the local pmcd service
and instead start the pcp-collector container, then run the same
tests again.

This test will _notrun if the 'atomic' or 'docker' commands
aren't available. It will also _notrun if the pcp containers are
not present.