pmdaroot: allow process creation without setting up fds
In the case of communication between pmcd and PMDAs, we defer
the setup of file descriptors until well after child process
creation (unlike in the case of pipes).  So, we need to allow
for this situation explicitly in the pmdaroot pmda-statup code
(i.e. its not an error to not send open file descriptors along
with a PID back to pmcd).

Problem uncovered by qa/110, additional diagnostics added into
pmdaroot that helped with figuring this out.