Alter certutils subject invocation and regex order
qa/common.secure - the 'common name' in the subject field of the
certutil invocation was being specified in such a manner that dropped
the quotes, thus, exposed the ',' character, which is used as a 'special
character' according to rfc1485. This was causing an 'improperly
formatted name' error on multiple platforms.