qa/common.avahi: capture stderr from avahi-browse
We were seeing failure text in .out.bad files like ...
    Failed to resolve service 'PCP pmproxy on foo.localdomain' of type '_pmproxy._tcp' in domain 'local': Timeout reached
where foo.localdomain was NOT the system under test, but was a system
that had recently been powered down and this seems to confuse avahi and
there is stale state preserved locally.

Since this has nothing to do with the PCP tests in question, capture
stderr from avahi-browse and salt it away in the .full file, but don't
pollute the .out file with messages like this.