qa/370: deal with arithmetic uncertainty
For reasons that remain a "Mystery, Berryl", there is some rare
differences in arithmetic between pmdumplog and sar2pcp, e.g. for
one metric at one point in time the value is 0.37 vs 0.38 similarly
there is another metric and instance and point in time where the
value maybe 0.00017 or 0.00018.

Add some fuzziness in the filtering to accommodate this.

A real fix would be nice, but that has so far escaped us.