Call Remove script when uninstalling single pmda's
With the current sub-rpm per pmda setup.  When an enabled pmda is
removed via rpm -e, pmcd.conf still contains the pmda line, and will
attempt (and fail) to restart the pmda on a pmcd restart.

Thus, adding a global pmda_remove macro to run the Remove script of a
pmda sub-rpm as it's erased.

RedHat BZ1304722