pcp-vmstat: wrapper for pmstat to be more vmstat-alike
Wrapper around pmstat to operate in a vmstat-alike way.
Test qa/804 added to exercise the script.

In the process found the use of proc_running & proc_blocked
to be needed, which we were not exporting from /proc/stat;
instead pmdaproc metrics were being used (which require a
/proc/*/stat scan and associated process state accounting in
the PMDA) for information the kernel already makes available
more efficiently.  Added the missing metrics into pmdalinux,
and start using them throughout.