qa: add unit tests for perfevent derived events
This commit adds unit tests for perf derived events.  The first test
case is to check the number of the perf derived events.  Second case
checks if there is a mismatch between the cpu configuration of the
perf derived events' counters.  Final test case checks if there is a
counter in derived list which is not enabled in the perfevent.conf
base list.

Signed-off-by: Hemant Kumar <hemant@linux.vnet.ibm.com>