pmrep: further work on pmrep signal handling warnings
This update cures all the cases for me, makes the code a bit easier
to read (IMHO), and still passes QA/880. The key difference is that we
don't touch stderr in finalize() at all, in a standalone test program
it works ok but errors out in pmrep although we don't do anything with
stderr in pmrep (so perhaps we're hitting some corner case of the Python
wrapper or libpcp itself). But since we now always flush stderr when
we write something into it I don't see a need to flush it again in
finalize(). Also flushes the output file a bit harder (as per os.html).