zbxpcp: support both Zabbix 2 and 3 versions
Recently released Zabbix 3 has introduced an API/ABI breaking
change in the module interface.  This commit switches between
use of one or the other API on the fly, using dlopen trickery
to detect the approriate ABI to use.  The header file allows
structures for both ABIs to co-exist and macros are shared.

It sounds neat but its all a bit of a hack.  Test qa/1067 has
good coverage of this functionality, so no additional test is
added here - just run with either zabbix version installed.