pcp.env: extension to optionally process $HOME/.pcp/pcp.conf
For some platforms, we rely on sudo(1) to install, start, stop
or remove PCP components, or to debug PMDAs that need special
privileges.

Unfortunately, sudo(1) may scrub the environment, meaning important
settings that are not in the system-wide pcp.conf (e.g. PERL5LIB
or PYTHONPATH) are cleared, resulting in script failure.

$HOME/.pcp/pcp.conf provides a way to overcome this as the scripts
of interest all source pcp.env at the start.

Not suprisingly, it is Mac OS X that has forced this "extension" so
that QA works at all for some tests.