qa/416: sudo and awk changes
The "sudo dbpmda ..." combination does not work well if the PMDA to
be run by dbpmda depends on environment variables that sudo(1) clears.
Leverage the recent pcp.env changes, and change this to:
    sudo sh -c ". /etc/pcp.env; dbpmda ..."

The version of awk on Mac OS X is pre-Neanderthal ... rework the
awk script to use day zero awk syntax.