qa assorted: more sudo problems, this time for dbpmda
The cause is sudo(1) clearing $TERM from the environment ... not much
of an issue, you say, but "$sudo dbpmda" prattles on for Mac OS X
because dbpmda uses readline which depends on terminfo which babbles
when TERM=unknown.

Replace dbpmda invocations like this with "$sudo TERM=ansi dbpmda ..."
or something similar.