configure & python pmapi wrapper changes: structs timeval and timespec
We cannot assume the type (size) of the fields of these structs
will be the same across all platforms.

Move to have the size determined by configure and then push these
into the configuration of the wrapper code.

This change makes QA 810, 842 and 859 all pass on Mac OS X (where
tv_usec in struct timeval is c_int, not c_long).