python api: reduce earlier timeval fix to just problem case
This is a simplified version of Kens earlier struct timeval
fix, which deals just with the problematic suseconds_t case.
This is handled more like the way differences in long/int/
pointer are handled now (in terms of configure - not python
specific), and more like the way the pmUnits structure is
conditionally setup (in terms of pmapi.py).

While researching alternative approaches, I came across the
base ctypes support for int32/uint32/int64/uint64 which we
should be using in pmAtomValue instead of guessing - so that
trivial type rectification is made here as well.