images: switch to DFSG-compliant licensing for most GUI icons
The Debian folk have pointed out a number of the icons used
in the Qt-based tools are using the old CC-SA-2.0 license and
that is now considered non-free.

Many of these icons originated from the Tango! project.  With
the v0.8.90 release of this iconset, all icons were placed in
the public domain.  So, for all icons that we did not modify,
or create ourselves, this commit switches to using the public
domain versions.  Some of the icons below images/ were also
discovered to be no-longer-used in the git tree, so these have
been removed entirely.

There remain a handful that were created/edited locally (by me)
that will need to be adjusted further - in a followup commit.