pcp-collectl: handle missing network.interface.in.compressed metric
Ken found qa/709 fails on MacOSX due to network metrics differences
on that platform when compared to Linux, leading to:

> ! Traceback (most recent call last):
> ! File "/usr/local/bin/pmcollectl", line 9, in <module>
> ! ssx.print_line()
> ! File "/usr/local/bin/pmcollectl", line 9, in print_line
> ! self.print_verbose()
> ! File "/usr/local/bin/pmcollectl", line 9, in print_verbose
> ! sum(self.ss.get_metric_value('network.interface.in.compressed')),
> ! TypeError: 'int' object is not iterable

He diagnosed root cause to:  "It appears that get_metric_value() is
returning a scalar value of 0 when values are not available for the
metric ... this is sort of a problem when the caller wraps this in
sum(...) because scalar 0 is not iterable.".

This commit provides a reworking of the extraction of these metrics.
Longer-term it would be good to move away from pmsubsys and its Linux
specific (arch-specific too with kernel.percpu.interrupts) hard-coded
metric lists, but this should resolve the QA failure in the interim.