RHBZ1325363: multithreaded pmNewContext
While parallelizing pmmgr, it was discovered that the core
pmNewContext function is a bottleneck when trying to connect to a
large number of servers.  Prior to this patch, it held the big libpcp
lock throughout the entire context-creation process, which can last
10+ seconds (e.g., if a remote pmcd host is unreachable).  That locks
out many other pmapi operations, and serializes connections to
multiple hosts.

Detailed analysis of pmNewContext and its callees showed that it is
possible to relax holding the big libpcp lock to much shorter time
periods, and specifically to exclude indefinite-length operations like
the socket connection to a remote pmcd, and even the analysis of
archives.  This is partly done by introducing a special
PM_CONTEXT_INIT c_type placeholder object into the context[] array
during initialization, and tweaking timing & locking sequences.

The result is that pmNewContext calls can almost completely overlap
each other safely.  A new test case (4751, a descendant of 475)
stress-tests by opening hundreds of various types of contexts at the
same time, including repeated, unreachable, and
theoretically-shareable ones.  The new code precludes sharing of
concurrent connections/archive-control data to the same destinations,
but leaves non-concurrent sharing behaviour is unmodified.