pmmgr: parallelize potential target pmcd analysis & daemon shutdown
It was reported that if pmmgr was given a target pmcd list containing
numerous hosts that are at times unreachable, then a delay of up to
$PMCD_CONNECT_TIMEOUT (10s!) may be absorbed - per unreachable host -
during the hostid-calculation phase.

So now we parallelize a couple of more things, to let pmmgr scale out
to a much larger number of target daemons:
- pcp contexts are opened in parallel to the potential pmcd list
  already gathered from target-host and target-discovery
- container subtargets are searched in parallel for surviving
  live pmcds
- eventually, pmmgr daemons are shut down in parallel, in separate
  threads that issue the SIGTERM / SIGKILL)

qa/666 updated.  Other scale testing with hundreds of
always-unreachable hosts (e.g., the RFC5737 TEST-NET 192.0.2.0/24
range) indicates proper parallelization and tolerance of timeouts.

Amongst some tasty coding treats:
- a "locker" class to embody automatic {}-block-lifespan mutex
  holding, instead of explicit pthread_mutex_[un]lock ops
- an "obatched" ostream-like class to let output-streaming <<
  operations accumulate in a stringstream, so concurrent cerr
  output is not interleaved
- a "parallel_do" function that launches N threads against a shared
  (usually embedded-lock-carrying) work-queue structure