pmmgr: tweak threading and verbosity
The recent pmcd-search multithreading work spun off threads up to a
calculated or configured limit, where that limit was independent of
the amount of work available for the threads.  This could waste time &
momentary memory.  We now limit multithreading to the actual number of
input work items.

While in the vicitinity, tweak message-verbosity so that pmmgr -v
prints a good bare-essential level of information (remote pmcds
found, daemons started), which is a good default.  -v -v prints
much more detail.