pmcd pmda auto-restart: pcpqa fallout
A few pcpqa tests specifically use crashy or timeout-inducing pmdas to
test pmcd's response to it.  With auto-restart, the response is
different, so these tests need updating.  Some of the tests are
essentially unmodified, just outputs regenerated (to assert
auto-restart), and 326 is changed to double-kill pmdas (to assert
auto-restart rate limiting).