pmdaproc: fix per-process scheduler metrics - type and semantics
Correct the type to signed-32bit-int and semantics to instantaneous
for the proc.psinfo.nice and proc.psinfo.priority metrics, after a
bit of kernel code review.  Also followed up on Markos point that
proc.psinfo.tty_pgrp looks dodgey - found the kernel can return -1
in the case that no tty is associated with the process.  So we will
return no-values-available in that case, instead of (unsigned)-1.

Log rewriting rules have been put in place, and qa/971 is added.

Resolves Red Hat BZ #1328432.