perfevent_pmda: Add capability for alternate event groups for derived events
This patch adds a capability to the perf derived events by giving them
a layer of abstraction. A perf derived event can be generated from a
group of base counters. The derived event can be a generic metric
whereas, the base counters are specific to hardware. For e.g., say a
metric called "memory_bandwidth" is defined as a derived metric. This
is obtained by looking at the base counters. These base counters will
be different for intel and other architectures. This patch gives the
perf derived events a layer of abstraction and ensures freedom in the
naming of the derived event. They can be defined in the perfevent.conf
file as :
...
[derived_perf_event:derived]
counter1_x86
counter2_x86
counter3_x86
counter4_x86
||
counter1_ppc
counter2_ppc
counter3_ppc
||
...

Note that the counters in derived group should have been defined in
perfevent.conf. The first group where all the counters are available
will be activated and rest won't be checked. If any counter in a group
is not available, entire group is discarded and the next group is
checked. Correct syntax has been specified in the perfevent.conf file.

Signed-off-by: Hemant Kumar <hemant@linux.vnet.ibm.com>