libpcp: fix brown paper bag pmcd-context-sharing bug
Rewrote test qa/246 (has had some problems since adding global
derived metrics support) to use valgrind instead of doing leak
detection "by-hand" using malloc pointer range checks.

This immediately showed a memory leak between pmNewContext and
pmDestroyContext, via the c_pmcd pointer when a few contexts are
in play.  Turns out to be due to a failure to perform reference
counting correctly, which in turn was due to a missing "break"
statement in the loop seeking a PMCD context sharing match.  It
is nasty not only because of the incorrect refcnts and leaks, but
because it basically disables sharing for the case of different
host contexts (only when the final context in the context array
happens to be a match would sharing actually happen correctly).

While doing root cause diagnosis, fixed up some whitespace damage
recently introduced into the libpcp context setup code too.