add grafana scripted dashboards for host-selection and queryparam-based graphs
Grafana "scripted" dashboards allow javascript code to generate
dashboard configurations on the fly.  See also;
http://docs.grafana.org/v1.9/reference/scripting/

The hostselect.js dashboard uses a brief pmapi/graphite ajax query to
deduce the list of host names (really, directories) with archives
accessible to pmwebd, and generates a little menu of hosts, from which
to pick.  It even unmangles the graphite archive/hostname paths.  It
is parametrized with url querystring fields for the set of metrics to
relay (&metric=) and styling options.  Each link relays those, and
navigates to a host-specific invocation of ...

The multichart.js dashboard generates a series of panels from graphite
metrics specified on the url querystring.  (Each metric is specified
by a &target=FOO.BAR pattern.)  A very limited set of configuration
options for each are also available. (&title=, &style={png,flot},
&height=).