build: add deb build dependency on manpages (configure.ac assumes it)
The ./configure checks for man page format (gz/bz2/xz/etc) sniff on
the local filesystem to determine which form is in use.  When .debs
are built in a chroot, its possible that becomes indeterminate.  So,
force the issue by ensuring base manpages package is there in chroot
builds too.  This is a possible root cause for this build failure:
https://github.com/performancecopilot/pcp/issues/8