pmwebd speedup: libmicrohttpd TURBO mode
An implementation artifact in libmicrohttpd prior to svn commit r37105
meant that concurrent requests into pmwebd are batched in the sense
that the response to one is not sent until the response to all are
finished.  This means more perceived waiting for e.g. pmwebd grafana
dashboards with multiple charts, because the empty screen lasts
longer.

The MHD_USE_EPOLL_TURBO flag for MHD_start_daemon activates
performance tweaks, including an improvement in the above behavior.
It's harmless in older libmicrohttpd, and is transparent to qa.