Allow the name of the server certificate to be customized
Add a -M option to pmproxy and pmcd to change the name of the
collector certificate that is searched for in the nssbd.

Also, when a server sends their certficate to the client.  The
original nickname is not sent.  Use the subject as the name to
save under.  Since under nss, the nickname:subject mapping must
be 1:1. If a client received certificates with different subjects
from different servers, the certificate save would fail with a
hardcoded nickname on the second certificate.