Update hostname for __pmSecureClientHandshake in pmproxy
I think this should be the peer hostname, not localhost. As far as
I can tell, all other uses of this function set it to the remote
pmcd hostname.  This is used in the call to SSL_SetURL/SSL_RevealURL
in order to validate the name in the remote cert.