pmmgr: tune logging batching
When pmmgr runs pmlogcheck on an archive, this can produce voluminous
warning traffic (e.g. for SGI PR1142) that's not helpful for a pmmgr
admin.  We now redirect that output also to /dev/null.  Since there is
now less output, tweak the obatched(stream) code to issue an explicit
ostream::flush(), so that whether the stream is default-buffered or
not, the log file will be current.