pmcd: make interrupted pmcd/pmda communications more resilient
Caught a pmcd shutdown during a QA test as a result of pmcd not
handling EINTR for agent communication (as it does for clients).

The specific failure observed was on fetch, but handling of the
store PDU had the same issue.  The latter also had a case where
it tested in incorrect error code, also fixed up here.