rc scripts: add primary (local) pmie concept to pmie
Similar to the primary pmlogger, this extends the pmie
control format to allow creation of a primary localhost
pmie process which enabled localhost-specific rules.

The stopped-pmda detection rule is added in as the first
localhost-specific rule, to improve the reliability of
the pmda restart process.

QA to follow, but this change is backward compatible in
control file format (same techniques that were used for
pmlogger are used here).

This resolves Red Hat BZ 1323851.