experimental libpcp multithreading alternative: private lock for contexts[]
We replace big-libpcp-lock usage to protect contexts[] with a
specialized lock just for that.  This means that the libpcp lock
becomes safe to acquire while holding a context c_lock.  This should
reduce the number of changes necessary for deadlock-proofness
elsewhere in libpcp.

The rest of the patch reverts several of the small-nestable-lock
changes from earlier in this branch.  The resulting code contains
fewer changes relative to the status-quo, but is not as robust.