packaging for OpenMandriva
This used to work a long time ago, but had not been tested for ages.

These changes get part of the way there, especially dealing with RPM
spec file differences and the package checks in check-vm, but more
needs to be done, especially since rpmlint is run by default here and
there are a boatload of rpmlint errors reported (many of which appear
to be rpmlint problems, but some are probably generic issues with the
PCP rpms on all platforms)  ... I've deferred this investigation
for another day.

Also refactoring of check-vm to greatly improve the performance in the
inner loop by not repeatedly checking for (and failing as expected)
the packaging tools that are not installed on the current platform.