Makepkgs: fix TAR setting
There had been some regression here, I suspect because depending on
the optional pieces being built, a valid setting for $TAR may or may
not be optional, because I was seeing:

    ./Makepkgs: 533: ./Makepkgs: -zxf: not found

'cause $TAR was undefined.

First round of changes here are to do a better job in gnu_tools()
of finding a gnu-style command (this is used to set $TAR).

May be some subsequent changes to actually fix the broken logic in
Makepkgs in the event that $TAR remains undefined and we try to use
this is the name of an executable gnu-style tar.