qa: rework python-not-configured checks for tests
Found several tests with python dependencies that we were
not checking for common.python with the original scheme.
Now, we unconditionally include common.python (instead of
compile-time conditional) so tests can always use it (and
are simpler as a result), and use the _notrun check there
for PCP_PYTHON_PROG as the main test guard.