src/pmie/pmie_check.sh: extra hardening
Greatly improved the error checking of the control file in the
wake of the great qa/115 disaster.

Also expunged the chown -R that is dangerous in the extreme ...
I don't think this is necessary at all once the directory containing
the log file and the log file itself (if it exists) are accounted for
with explicit chown actions.