pmdammv, qa: hardening mmap file data handling
Numerous updates to pmdammv to make sure it cannot be coerced
into accessing memory outside of the memory mapped files that
it loads.  In addition, sensible (but still large) limits are
placed on the data structures that are allocated to hold data
representing various parts of the mappings.

The mmvdump utility is updated in a similar way to improve its
resilience in the presence of bad mappings.  It also has some
slight output refinements (symbolic flag printing, etc).

This tackles SGI BZ 1062.