libpcp: change a derived metric warning to pmDebug diagnostic
When pmlogger records a derived metric, and then a client tool
replays with the same derived metric config active, a warning
(or series of warnings, one per metric) is generated.  This is
a normal, expected situation though when global derived metrics
are being used.

Added a slightly modified version of Marks test case as qa/1039.