pmdaproc: rewrite proc.psinfo.ttyname metric handling
Resolves some short-comings in proc.psinfo.ttyname
- stat'ing and realpath'ing every open fd when a dev match
  not found can be expensive;
- it also generates many SElinux warnings, since the fds
  can point anywhere (or maybe sockets, fifo's, etc);

Restrict the search to just the first three (standard) fds
(even though these may be remapped) since its not likely
we will find a match if not here.  Use readlink on the proc
files instead of stat to avoid following the link, keeping
SElinux happy.  Only stat links that begin with /dev/ (the
code assumed this path previously anyway, so check for that
before making any further syscall).

Related to RH BZ #1336211.