pmlogger: auto-reconnect to pmcd whenever possible
Extend Dave's consistency checking to also allow pmlogger
to reconnect automatically to pmcd (just like pmie always
has).

This is important for the common case of a pmcd restart on
a host monitored by a central pmlogger farm.  Previously,
we were open to some minutes of missed data until the cron
pmlogger_check kicked in - now it is instantaneous.

Its worth noting that we still always start a new pmlogger
with each new day, even with this change in place.

Verify the auto-reconnection and presence of a mark record
via the qa/389 test script.