python api: drop the incomplete gui mode support interfaces
As discussed in the BZ below, this commit removes the currently
incomplete GUI mode support from the Python PMAPI.  Completing it
would require a non-trivial amount of work, there's been no strong
need for it to be finished, and the current pmapi.py gives false
image it being usable. Of course, if someone wants to work on this
in the future then the effort can be started by reverting this.

https://bugzilla.redhat.com/show_bug.cgi?id=1289910