pcp2influxdb: export PCP metrics to InfluxDB
We are working on switching from graphite to InfluxDB for metric
storage, mostly for performance reasons.  In order to do this, I
copied pcp2graphite and did a bit of a re-write to get it sending
metrics to InfluxDB.

InfluxDB has an HTTP API, which is the main way of sending data
to it.  This patch handles cases with and without authentication.