src/pmlogger/pmnewlog.sh: improve checking
pmnewlog is supposed to restart _exactly_ one pmlogger process.

If it find more than one "target" processes it should report and quit.
This situation is most unlikely in production environments but has
been seen in QA, with some earlier test probably causing qa/1108 to
fail as a result ... this change should caue qa/1108 to fail with
more useful error messages.