src/pmcd/src/client.c: small change to new client connection logic
In AcceptNewClient(), ECONNABORTED is a (sort of) expected result
from __pmAccept() ... at least it has been seen in QA on FreeBSD
platforms.

Also in the same area, if the error from __pmAccept() is not
expected we should just drop the client connection, not terminate
pmcd as was previously the case.