libvirt pmda: support old libvirt 0.10 API
Add support for additional metrics available using the old and
clumsy API.

The code certainly is not pretty but I can't see how it could be
radically cleaner. The good thing is that it's well-contained and
doesn't interfere with the code using the newer API.

This is mostly for RHEL 6 / libvirt-0.10 environments but since now
configurable, can be enabled anywhere where the newer API is not
an option for some reason.