qa/pmdas/broken/broken_pmda.c: try to avoid qa/274 race
For the BUG_5 version, the PMDA will quit in pmdaDaemon() which
introduces a possible race with dbpma sending a PDU to a process
that is about to exit, but has not yet closed the IPC channel.

Explicitly fclose()ing stdin and stdout before exit() seems to make
qa/274 more robust.