RHBZ 1348351: pmchart(1): Detect archives from the same host.
The archives specified on separate -a options must be from different hosts.
pmchart(1) was failing to detect this. Failure to detect this results
in only the first archive for each host being visible to pmchart(1).
This was the cause of the errors being reported in the bug report.

pmchart(1) will now issue an error message reporting the duplicate
hosts and suggest combining them using pmlogextract(1).