qa/870: (new) test integrity of pmlogger control files
These are the ones in $PCP_TMP_DIR/pmlogger.  And getting this test
to pass will address the root cause of the non-deterministic qa/1108
failures.

This test can be run with a --check argument which silently (if all
is well) runs the integrity check without any of the test cases.
In this form, could be used with check.callback to run the check
after every test to help identify any test that leaves the control
files in a bad state.