src/pmlogger/src/ports.c: fix broken logic for primary control file
This was the root cause of the qa/1108 failures.

The logic that checked for and stopped more than one primary pmlogger
from running was broken.  Specifically using stat() instead of
lstat() to check for a symbolic link will always fail, which drove
us down the "old-style hardlink" path and unconditionally removed
$PCP_TMP_DIR/pmlogger/primary before the existance check that was
intended to stop multiple primary loggers from running.

This error seems to have been introduced in commit 7148bf11 (almost
12 months ago) ... sigh.

And to compound the problem, a primary pmlogger was conditionally
removing $PCP_TMP_DIR/pmlogger/primary at exit, meaning that if we
ever got 2 (or more!) primary pmloggers running and either of them
exited the control files would be removed and pmlogger_check would
stumble along later and start another primary pmlogger running.

So now we are checking the pid from the symlink and only removing
the primary control file if this instance of pmlogger created it.

Also cleaned up some misleading diagnostics.