zsh compilations: -w is not needed for _arguments
-w was unneeded and in fact incorrect:

The  combination  of  -s  with  -w  allows  single-letter
options  to  be  combined in a single word even if one or
more of the options take arguments.  For example,  if  -a
takes  an  argument,  with no -s `-ab' is considered as a
single (unhandled) option; with -s -ab is an option  with
the  argument  `b';  with  both -s and -w, -ab may be the
option -a and the option -b with arguments still to come.

That is, with -w, -xy <tab> arguments for -x would be completed.