qa/check: extend check.callback functionality
Previously if check.callback existed and was executable in the
directory that check is run from, then check.callback was run
after _every_ qa test.

Extend this so that if check.callback returns a non-zero exit
status all subsequent QA tests are abandoned.

This is most useful in locating pathological problems that are
caused by one test, but not noticed until some (maybe much) later
QA test fails.  For example, this allowed qa/964 to be identified
as one of the causes for qa/1108 failing sometimes.