qa/381: at last, problem identified
This test has been failing intermittently for some time and I've
finally tracked it down ... the logic assumed that the primary logger
was handling pmlc requests on port 4330.  This is true most of the
time (test passes) but sometimes (especially with pmmgr in the mix)
may not be so (test fails).