RHBZ 1351223: multi-archive mode unnecessarily asserts matching loglabel timezones
It is not necessary to reject archives created in different time zones
in a multi-archive context. Changes to and from daylight savings
time are examples of such differences. Even pmlogextract(1) tolerates
such differences and chooses either the first or the last time zone
seen as the time zone for the combined archives.

This change removes the check for archives with different time zones
in a multi-archive context.