src/libpcp: make the pmcd-client protocol safer in the presence of timeouts
If a timeout occurs on a PDU read (in __pmGetPDU) the libpcp callers
are now responsible for closing the channel that connects pmcd and all
contexts associated with that pmcd in the current client app.

Because this channel maybe multiplexed between contexts, a timeout
introduces the possibility that incorrect PDUs are seen by other
contexts and this breaks the synchronous communication between pmcd
and a PMAPI context, or worse, delivers incorrect data to other PMAPI
contexts.

This change prevents that from happening, but the other contexts
will need to call pmReconnectContext() before they can resume sending
and receiving PDUs to and from the pmcd that triggered the original
timeout condition.