pmchart(1): Change in data structures for chart data.
QwtPlotCurve::setRawSamples is deprecated and we needed to
switch to another method for setting the plot data. This suggested
switching the plot data from array of double to a vector of something.

The initial idea was to switch the data to QVector<double> and to use
QwtPlotCurve::setSamples( const QVector<double> &xData,
const QVector<double> &yData ). However it looks like this
method is also on the way out. The final implementation keeps
the data as QVector<double> and constructs a QVector<QPointF>
so that QwtPlotCurve::setSamples( const QVector<QPointF> &samples )
can be used.